+++ plugins/tech-insights-backend/src/service/fact/factRetrievers/entityOwnershipFactRetriever.ts
@@ -14,12 +14,10 @@
  * limitations under the License.
  */
 
-import {
-  FactRetriever,
-  FactRetrieverContext,
-} from '@backstage/plugin-tech-insights-node';
+import { FactRetrieverContext } from '@backstage/plugin-tech-insights-node';
 import { CatalogClient } from '@backstage/catalog-client';
 import { Entity } from '@backstage/catalog-model';
+import { FactRetriever } from '@backstage/plugin-tech-insights-common';
 
 /**
  * Generates facts which indicate the quality of data in the spec.owner field.
