+++ plugins/tech-insights-backend/src/service/fact/factRetrievers/techdocsFactRetriever.ts
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-import {
-  FactRetriever,
-  FactRetrieverContext,
-} from '@backstage/plugin-tech-insights-node';
+import { FactRetrieverContext } from '@backstage/plugin-tech-insights-node';
 import { CatalogClient } from '@backstage/catalog-client';
 import { Entity } from '@backstage/catalog-model';
 import { entityHasAnnotation, generateAnnotationFactName } from './utils';
+import { FactRetriever } from '@backstage/plugin-tech-insights-common';
 
 const techdocsAnnotation = 'backstage.io/techdocs-ref';
 const techdocsAnnotationFactName =
