+++ plugins/tech-insights-common/package.json
@@ -35,7 +35,8 @@
   "dependencies": {
     "@types/luxon": "^2.0.5",
     "luxon": "^2.0.2",
-    "@backstage/types": "^1.0.0"
+    "@backstage/types": "^1.0.0",
+    "@backstage/plugin-tech-insights-node": "^0.3.1"
   },
   "devDependencies": {
     "@backstage/cli": "^0.18.0-next.1"
